requirements.txt
beautifulsoup4==4.8.1
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
google-api-core==1.14.3
google-auth==1.7.0
google-cloud-vision==0.39.0
googleapis-common-protos==1.6.0
grpcio==1.25.0
gTTS==2.0.4
gTTS-token==1.1.3
idna==2.8
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
PyAudio==0.2.11
pyfiglet==0.8.post1
pygame==1.9.6
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
soupsieve==1.9.5
SpeechRecognition==3.8.1
urllib3==1.25.6
